Sample showing potential angular2/web component bug
Home page now has a google-chart web component, initially displaying a
pie chart. Clicking the Refresh button should change it to a geo chart.
It does not.

However, the “Trial” link has exactly the same code and google-chart.
On this page, if you click Refresh the chart visually changes.

Note that if you are on the Trial page and refresh the web page, this
chart will stop updating when you click Refresh. However, if you then
click Home and the Refresh button, that chart will then redraw properly.